chore(wmg): move Loader and ScreenTint
