+++ frontend/src/views/WheresMyGeneV2/components/ScreenTint/index.tsx
